Pitch-class chord alphabet

This dataset defines an alphabet for the type "pc_chord". It is a list with two components, defining a forward and backward mapping between integers and "pc_chord" objects:

by_id: A list object. The ith element of this list contains the pitch-class chord corresponding to integer i.

by_chord: An environment object. A pitch-class chord can be mapped to an integer by taking the pitch-class chord object, converting it to a string representation (with as.character), and accessing the object within this environment with this name. The result will be a coded vector (see coded_vec) of length 1.