pc_chord to pc_set map

This integer vector provides the mapping between pitch-class chords ("pc_chord" objects) and pitch-class sets ("pc_set" objects). This can help computational efficiency in some situations.

Details

If i is the integer corresponding to an encoded pitch-class chord, then the corresponding encoded pitch-class set is equal to the ith element of this vector.

Examples

i <- encode(pc_chord(c(4, 0, 7)))
j <- pc_chord_id_to_pc_set_id_map[i]
decode(coded_vec(j, "pc_set"))[[1]]

#> Pitch-class set: 0 4 7