Smooth pitch spectrum

Creates an object of class smooth_pi_spectrum, describing a smooth pitch spectrum. A smooth pitch spectrum describes perceptual weight along a continuous MIDI pitch scale. The representation incorporates smoothing to account for imprecisions in pitch perception.

smooth_pi_spectrum(x, sigma = 6.83, ...)

# S3 method for default
smooth_pi_spectrum(x, sigma = 6.83,
  num_harmonics = 11L, roll_off = 1, ...)

# S3 method for sparse_pi_spectrum
smooth_pi_spectrum(x, sigma = 6.83, ...)

Arguments

x
Object to convert. By default sparse_pi_spectrum is called first to convert the object to a sparse pitch spectrum.

sigma
(Numeric scalar) Standard deviation of the Gaussian distribution used to simulate perceptual blurring. Defaults to 6.83 cents, after Milne and Holland (2016) .

...
Provided for S3 method consistency.

num_harmonics
(Integerish scalar) Number of harmonics (including the fundamental) to which each tone should be expanded.

roll_off
(Numeric scalar) Parametrises the amount of amplitude roll-off in the harmonics, with greater values corresponding to higher roll-off.

Details

This representation is created by expressing the sonority as a sparse pitch spectrum (see sparse_pi_spectrum) and convolving the result with a Gaussian distribution with standard deviation sigma.

See also