Sparse pitch spectrum

This function represents an input object as a sparse pitch spectrum.

sparse_pi_spectrum(x, ...)

# S3 method for fr_sparse_spectrum
sparse_pi_spectrum(x, ...)

# S3 method for list
sparse_pi_spectrum(x, ...)

# S3 method for default
sparse_pi_spectrum(x, ...)

# S3 method for pi_chord
sparse_pi_spectrum(x, amplitude = 1, ...)

# S3 method for sparse_pi_spectrum
c(..., x_digits = 6)

Arguments

x
Input sonority.

...
Further arguments passed to expand_harmonics(), depending on the method invoked.

amplitude
(Numeric vector) Vector of amplitudes to assign to each pitch. If a scalar value is provided, this value is assigned to all pitches

x_digits
(Integerish scalar) Number of significant digits to which pitches are rounded.

Value

An object of class sparse_pi_spectrum.

Details

A sparse pitch spectrum describes an input sonority as a finite set of spectral components, each defined by a pitch (expressed on the MIDI pitch scale) and an amplitude (expressed in arbitrary units, but with the fundamental frequencies of chord pitches typically taking the value 1).

Sparse pitch spectra can be combined into one spectrum using c(...). Amplitudes are summed assuming incoherent wave superposition (see sum_amplitudes).